Workflow autoloading; button outlines; bug button visibility.
Closes #7.
Added outline around edit and add-new buttons in collection table to make them more obvious targets.
Added a preference in the main Settings app to hide or show the raw output button (the bug button). By default, the button will be hidden.
Any time a workflow is edited (if there are no collections) or created, it's now autoloaded into the detail view controller.